pyproject.toml
[project]
name = "rag-ia-agent-portuguese-opportunities"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.11",
    "langchain-google-genai>=4.2.1",
    "langchain-postgres>=0.0.17",
    "langgraph>=1.1.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
